prepares our macro defs for refactoring (Step I)
At the moment macro definitions get detected as such during parsing,
when their right-hand sides are found to begin with the "macro" keyword.

We intend to refactor this syntax, changing it from "macro <macro impl>"
to "macro(<macro impl>)" and dropping the keywordness of "macro", turning
it into a magic method Predef.macro. The motivation for that is to free up
the keyword to rename the "scala.reflect.makro" package.

This refactoring conflicts with preexisting macros defined in our codebase,
because some of them (namely, tag materialization macros and string context
"f" formatter) are used inside the compiler itself.

The logic of those macros is hardwired into starr's fast track,
so it doesn't rely on any of the subsystems of the macro engine
to be located, bound and executed.

But to trigger this logic we need to color these macros as macros, i.e. as
term symbols having the MACRO flag. Since the parser is going to be changed
and will be no longer able to recognize macros as macros, we need to help
the compiler by explicitly setting the MACRO flag for the corresponding symbols