removes the old "macro <impl method>" syntax
From now on until the end of refactoring compiler doesn't recognize
any macros except for those that are hardcoded in the fast track.

Therefore all the macro tests will fail until the new macro syntax is in place.
As usual I prefer commit chain to be as clear as possible rather than pass all
the tests all the time.

I've also removed  -Xmacro-primary-classpath and -Xmacro-fallback-classpath.
These facilities require macro definitions to be typechecked in Namers
(the exact reason can be found in comments to the deleted code in Namers).
Previously that was viable, because even before Typers we could distinguish
a macro from a non-macro. Now it's a hassle, so I'm just dropping this
functionality, since no one is using it anyways.

And by the way - "macro" is no longer a keyword.