adds the new "macro(<impl method>)" syntax
Now macro definitions are detected as being such when having bodies
that consist of a sole call to Predef.macro with a single argument
that represents a reference to a macro implementation method
(either unqualified or qualified, just as it was before).

In comparison with the amount of code removed in previous commit,
the changes are surprisingly miniscule and mostly consist of
renaming local variables in typedMacroBody in Macros.scala.