prepares our macro defs for refactoring (Step II)
Since the previous commit the compiler doesn't need any help
to find out that fast-tracked macros are macros. Hence I'm
replacing their RHSes with the universal ??? body.

This way macro definitions defined in scala (for example,
tag materialization macros or certain string formatters)
won't stand in the way of the planned change to the macro system.